make some components use reusable components
